# apps/web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ilios API",
    "description": "Local-first API for Ilios.",
    "version": "0.2.1"
  },
  "paths": {
...
    "/libraries/{library_id}/imports/arxiv": {
      "post": {
        "tags": [
          "imports"
        ],
        "summary": "Import Arxiv",
        "operationId": "import_arxiv_libraries__library_id__imports_arxiv_post",
        "parameters": [
          {
            "name": "library_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Library Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportArxivRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/libraries/{library_id}/imports/file": {
      "post": {
        "tags": [
          "imports"
        ],
        "summary": "Import File",
        "operationId": "import_file_libraries__library_id__imports_file_post",
        "parameters": [
          {
            "name": "library_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Library Id"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/ImportLocalKind"
            }
          },
          {
            "name": "file_name",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "File Name"
            }
          },
          {
            "name": "parse_after_import",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Parse After Import"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportLocalResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/libraries/{library_id}/articles/{revision_id}/cards/extract/jobs": {
      "post": {
        "tags": [
          "articles"
        ],
        "summary": "Post Reader Card Extraction Job",
        "operationId": "post_reader_card_extraction_job_libraries__library_id__articles__revision_id__cards_extract_jobs_post",
        "parameters": [
          {
            "name": "library_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Library Id"
            }
          },
          {
            "name": "revision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Revision Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReaderCardExtractionRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/libraries/{library_id}/articles/{revision_id}/cards/generate/jobs": {
      "post": {
        "tags": [
          "articles"
        ],
        "summary": "Post Reader Card Generation Job",
        "operationId": "post_reader_card_generation_job_libraries__library_id__articles__revision_id__cards_generate_jobs_post",
        "parameters": [
          {
            "name": "library_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Library Id"
            }
          },
          {
            "name": "revision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Revision Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReaderCardGenerationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/libraries/{library_id}/articles/{revision_id}/embeddings/jobs": {
      "post": {
        "tags": [
          "articles"
        ],
        "summary": "Post Article Embedding Job",
        "operationId": "post_article_embedding_job_libraries__library_id__articles__revision_id__embeddings_jobs_post",
        "parameters": [
          {
            "name": "library_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Library Id"
            }
          },
          {
            "name": "revision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Revision Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmbedArticleRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/libraries/{library_id}/articles/{revision_id}/exports/jobs": {
      "post": {
        "tags": [
          "articles"
        ],
        "summary": "Post Article Export Job",
        "operationId": "post_article_export_job_libraries__library_id__articles__revision_id__exports_jobs_post",
        "parameters": [
          {
            "name": "library_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Library Id"
            }
          },
          {
            "name": "revision_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Revision Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArticleExportRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/jobs": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Jobs",
        "operationId": "get_jobs_jobs_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 500,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Job"
                  },
                  "title": "Response Get Jobs Jobs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "jobs"
        ],
        "summary": "Clear All Jobs",
        "operationId": "clear_all_jobs_jobs_delete",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobClearResult"
                }
              }
            }
          }
        }
      }
    },
...
    "/jobs/{job_id}": {
      "get": {
        "tags": [
          "jobs"
        ],
        "summary": "Get Job By Id",
        "operationId": "get_job_by_id_jobs__job_id__get",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}/pause": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Pause Job By Id",
        "operationId": "pause_job_by_id_jobs__job_id__pause_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}/resume": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Resume Job By Id",
        "operationId": "resume_job_by_id_jobs__job_id__resume_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/jobs/{job_id}/cancel": {
      "post": {
        "tags": [
          "jobs"
        ],
        "summary": "Cancel Job By Id",
        "operationId": "cancel_job_by_id_jobs__job_id__cancel_post",
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArticleExportKind": {
        "type": "string",
        "enum": [
          "source_markdown",
          "translated_markdown",
          "bilingual_markdown",
          "lecture_notes",
          "bundle_zip"
        ],
        "title": "ArticleExportKind"
      },
      "ArticleExportRequest": {
        "properties": {
          "kind": {
            "$ref": "#/components/schemas/ArticleExportKind",
            "default": "bilingual_markdown"
          },
          "target_language": {
            "type": "string",
            "maxLength": 40,
            "minLength": 2,
            "title": "Target Language",
            "default": "zh-CN"
          },
          "include_untranslated": {
            "type": "boolean",
            "title": "Include Untranslated",
            "default": true
          }
        },
        "type": "object",
        "title": "ArticleExportRequest"
      },
...
      "EmbedArticleRequest": {
        "properties": {
          "provider": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "Provider",
            "default": "local-hash"
          },
          "model": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Model",
            "default": "hashing-64-v1"
          },
          "force": {
            "type": "boolean",
            "title": "Force",
            "default": false
          }
        },
        "type": "object",
        "title": "EmbedArticleRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ImportArxivRequest": {
        "properties": {
          "arxiv_id": {
            "type": "string",
            "minLength": 1,
            "title": "Arxiv Id"
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          },
          "download_pdf": {
            "type": "boolean",
            "title": "Download Pdf",
            "default": true
          },
          "parse_after_import": {
            "type": "boolean",
            "title": "Parse After Import",
            "default": true
          }
        },
        "type": "object",
        "required": [
          "arxiv_id"
        ],
        "title": "ImportArxivRequest"
      },
      "ImportLocalKind": {
        "type": "string",
        "enum": [
          "tex_archive",
          "markdown",
          "pdf"
        ],
        "title": "ImportLocalKind"
      },
      "ImportLocalResult": {
        "properties": {
          "library_id": {
            "type": "string",
            "title": "Library Id"
          },
          "article_family_id": {
            "type": "string",
            "title": "Article Family Id"
          },
          "article_revision_id": {
            "type": "string",
            "title": "Article Revision Id"
          },
          "bundle_path": {
            "type": "string",
            "title": "Bundle Path"
          },
          "source_kind": {
            "$ref": "#/components/schemas/ImportLocalKind"
          },
          "parse_job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parse Job Id"
          }
        },
        "type": "object",
        "required": [
          "library_id",
          "article_family_id",
          "article_revision_id",
          "bundle_path",
          "source_kind"
        ],
        "title": "ImportLocalResult"
      },
      "Job": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "type": {
            "$ref": "#/components/schemas/JobType"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "priority": {
            "type": "integer",
            "title": "Priority",
            "default": 0
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "error": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "progress": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Progress",
            "default": 0.0
          },
          "attempts": {
            "type": "integer",
            "title": "Attempts",
            "default": 0
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "lease_owner": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lease Owner"
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "Job"
      },
      "JobClearResult": {
        "properties": {
          "cleared": {
            "type": "integer",
            "title": "Cleared"
          }
        },
        "type": "object",
        "required": [
          "cleared"
        ],
        "title": "JobClearResult"
      },
      "JobStatus": {
        "type": "string",
        "enum": [
          "queued",
          "running",
          "paused",
          "succeeded",
          "failed",
          "cancelled"
        ],
        "title": "JobStatus"
      },
...
      "JobType": {
        "type": "string",
        "enum": [
          "import_arxiv",
          "parse_article",
          "translate_block",
          "embed_article",
          "export_article",
          "extract_reader_cards",
          "generate_reader_card"
        ],
        "title": "JobType"
      },
...
      "ReaderCardExtractionRequest": {
        "properties": {
          "target_language": {
            "type": "string",
            "maxLength": 40,
            "minLength": 2,
            "title": "Target Language",
            "default": "zh-CN"
          },
          "limit": {
            "type": "integer",
            "maximum": 120.0,
            "minimum": 1.0,
            "title": "Limit",
            "default": 30
          },
          "force": {
            "type": "boolean",
            "title": "Force",
            "default": false
          }
        },
        "type": "object",
        "title": "ReaderCardExtractionRequest"
      },
...
      "ReaderCardGenerationRequest": {
        "properties": {
          "anchor_block_uid": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Anchor Block Uid"
          },
          "anchor_text": {
            "type": "string",
            "maxLength": 240,
            "minLength": 1,
            "title": "Anchor Text"
          },
          "target_language": {
            "type": "string",
            "maxLength": 40,
            "minLength": 2,
            "title": "Target Language",
            "default": "zh-CN"
          },
          "provider_profile_id": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Profile Id"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "native_search": {
            "type": "boolean",
            "title": "Native Search",
            "default": true
          },
          "card_type": {
            "$ref": "#/components/schemas/ReaderCardType",
            "default": "term"
          },
          "title": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 240
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "abbreviation": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Abbreviation"
          },
          "full_form": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 240
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Form"
          }
        },
        "type": "object",
        "required": [
          "anchor_block_uid",
          "anchor_text"
        ],
        "title": "ReaderCardGenerationRequest"
      },
...
      "ReaderCardType": {
        "type": "string",
        "enum": [
          "term",
          "note",
          "formula",
          "figure",
          "question"
        ],
        "title": "ReaderCardType"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}